Added skeleton of SpriteMaker
Starting with this since it involves good LCD and input usage and
should allow some content creation.

Concept is to blit a grid to the LCD, the user can select a cell in the
grid using the sliders (select cell can flash on off). A cell is
filled/colored using a short button press or holding both sliders.

The state should track and store the coordinates of each filled cell.
This way we can redraw the sprite at various sizes easily, making this
a nice way to let the user create an avatar that we can use throughout the
badge modes.