Sub-bullets; Minor correction regarding state struct
Some of the original bullet-points were sub-bullets since they were in
regards to the parent bullet-point. Also, the state struct does have a
next state struct. Potentially a next state handler could be added, but I'm
not sure the performance tradeoff of having the extra assignment in the
critical path is worth the gain in elegance. Open to discussion and/or
testing.